requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
click==7.1.2              # via typer
numpy==1.19.3             # via lenny_intensifies (setup.py), scipy
pillow==8.0.1             # via lenny_intensifies (setup.py)
scipy==1.5.3              # via lenny_intensifies (setup.py)
typer==0.3.2              # via lenny_intensifies (setup.py)
